Organizaçao dos exercicios em diretorios e arquivos
