+++ exercises/modulo_1/bloco_7/dia_3/parte1/ex5.js
@@ -1,0 +1,18 @@
+const assert = require('assert');
+
+const obj1 = {
+  title: 'My Title',
+  description: 'My Description',
+};
+
+const obj2 = {
+  title: 'My Title',
+  description: 'My Description',
+};
+
+const obj3 = {
+  title: 'My Different Title',
+  description: 'My Description',
+};
+
+// Testes
